refactor dcp1 to dcp2 migration tool (#165) (#166)
* refactor dcp1 to dcp2 migration tool (#165)

The format on s3 for dcp metadata has changed.  so the tool needed to be updated.
For example, the object names do not indicate the entity type.  This can be determined
by processing the text of the file.

The whole script has been refactored:
  1. There was code to upload data to s3.  This has been removed.
  2. A new feature to read the metadata from a tar.gz file has been implemented.  It can
     read the tar.gz file inplace (without having to extract the contents as files).
  3. The script can first download the .tar.gz file from s3 and then process it.  i

Add a new script called metadata_aggregator.py:
  1. downloads all the metadata for a project from s3
  2. tar/gz the contents
  3. push the tar.gz file back up to s3

All 29 projects have been processed with the metadata_aggregator.py script.
Downloading the metadata from s3 and reading into the script now only takes a
few seconds instead of 10s of minutes.