332 cookie banner (#511)
This PR adds the cookie banner! More notes are added as comments 😄 

The pending decision here is whether to use cookie or localStorage to save the user preference, waiting on infosec's advice. The PR has commits for both, so we can switch to the recommended one once we get an answer!

Thank you 🎉 !

closes #332 

Mock:

<img width="1027" alt="Screen Shot 2020-08-06 at 9 35 50 AM" src="https://user-images.githubusercontent.com/6309723/89557855-505e2680-d7c8-11ea-8013-6c8cf3fd8022.png">

Real:

<img width="1366" alt="Screen Shot 2020-08-06 at 9 35 36 AM" src="https://user-images.githubusercontent.com/6309723/89557887-5bb15200-d7c8-11ea-9b72-895dd89542b2.png">