Change dataset ontology types to JSON (#542)
Datasets can have multiple values for things like tissue and assay. This changes those types from string to JSON and merges the ontology term id field with the label field.

Note that downgrading the database after this change is lossy, so if we decided to do that, there is some additional data migration work needed.