Haromize dev, staging, prod db schemas
This is an ugly commit, but the schemas of dev, staging, and prod had
drifted apart, weren't properly under the control of alembic, and
differed from the ORM code. This uses some autogenerated revisions
applied to each brach to bring them in line with the ORM code. The
alembic version 9023d0bf61ab was stamped on every environment's db, and
the histories were merged (kind of) via alembic merging.

There were a couple little raw SQL tweaks needed (time with vs without
time zone, names but not content of foreign key constraints, etc) in
addition, but now all three environments have schemas that are identical
per apgdiff and are on the same alembic version num.

One caution is that if you want to downgrade gracefully, you're going to
have a really bad time. But that was true before this commit too.