Getting 500 errors for GET Collection (#777)
* Getting 500 errors for GET Collection

When reading a dataset already in the database, not all fields expected according to the openapi spec were present. Since the only way to add dataset to a collection is through the upload of an H5AD file, which does its own validation, we can relax the fields required in the openapi spec. The spec has been modified to allow more field in the dataset to be null. This presently only affect responses we return since users cannot population these field through the API.

* allow for null values in the dataset metadata

Co-authored-by: Seve Badajoz <sbadajoz@chanzuckerberg.com>