Override config from env by default, functional-test reads from chamber (#980)
* FUNCTIONALITY CHANGE: Makes data portal read config from environment by default,
overriding values in config file. Uses existing functionality with minor code changes.
* Fix code and unit tests that were broken when use_env=True
* Fix non-deterministic behavior when reading config
* Unset variables that are overrridden by EnvironmentSetup decorator if not previously set
Explicitly set use_env in tests that were relying on implicit behavior.
* Uses chamber to read environment variables for auth0 for client ID and secret
when running make local-functional-test. This is a stopgap that will unblock
local testing while we work on a longer term solution. This has been tested
to pass tests successfully, and is currently hard coded to using the secrets
from the dev stage. Requires Chamber v2.9.0 or above to work
(needs secretsmanager backend introduced in that version).
This can be installed via brew install on Macs.