Handle workspaces without any latest config version
The happy path script breaks if a previous run created a stack's workspace
without actually creating any workspace state or workspace config versions,
and you try to delete that workspace.

This PR prevents us from crashing in these cases, instead printing a warning
that the workspace is empty. Since the workspace is empty, this should not
be harmful.