Pass git commit at container build time, use a next.js build id
* Make the current Git branch, commit hash, and tag description be
  added as Docker build-time arguments.
* Modify containers to add those as labels embedded within the image
* Override next.js build ID to use the commit hash, falling back to
  default random ID behavior if commit hash not provided.

Ideally we'd rather build the next build artifact right into the
container image, so a 2nd container using the same image would have
the same build ID for free (and there's no build time at startup).
However, we cannot run `npm run build` on the bare image because
even running this requires a configs/config.js file to be present,
which it isn't at the time we initially build the Dockerfile.