fix: Improve local dev compatibilities (#1732)
This commit makes some of the local dev commands more flexible to allow
'rdev' as an acceptable DEPLOYMENT_STAGE variable (for the purposes of
using an rdev database during local development). It also preserves env
variables for sudo commands in local dev Make targets which is necessary
in order to be compatible with some developers' virtual environment
configurations.