fix: detect duplicate H5AD files (#1784)
* fix: detect duplicate H5AD files

Only check h5ad File for duplication across datasets in a collection. Don't check derived files for duplication. This was added because smoke tests were failing with a 500 error when publishing a collection. This was caused by the way were are retrieving metadata from S3. When parsing the file name from the S3 URI we were not taking into account that the URI for the CXG artifact was to a directory. Thus when requesting metadata for a directory, S3 return 404. Rather than retrieving all of the files from the CXG directory and check for duplication it's faster and easier to check if the source H5AD file is duplicated across the datasets in the collection. Since the H5AD file is used to generate the other artifacts, it's safe to assume if the H5AD file are unique, then the derived artifacts must also be unique.