fix: update API response schema for Datasets (#2779)
* fix: allow processing statuses to be null

Apparently there exists some data corruption. Allow processing status
entity to be null in response for flexibility. Handle key error.

* fix logic

* remove processing_status from preview

This commit also addresses the issue of null processing status entities
in the dict version of the response

* make curator_name nullable in API def

* ensure that organism attr of response is an array

* match db schema definition for nullable attrs

The following Collection attributes are made nullable in the API
response:
- name
- contact_name
- contact_email

* fix default/array packaging issues with all ontology_elements

* manually assign empty array as default

The default value from Connexion is not kicking in because the attrs are
defined on the response object; they simply have None as their value.