fix: Fix rollup bug when selected tissue has no gene expressions (#4869)
* fix: Fix rollup bug when selected tissue has no gene expressions

When the query API gets a list of tissues and genes, and it is
the case that one of the tissues in the query arguments does not
express any of the genes in the query argument, the rollup function
would throw a KeyError when looking up the "gene expression free"
tissue in one of the data structures.

The fix is to prevent this look up by filtering these
"expression free tissues".

* Check key in dictionary in a pythonic way