Create a V2 version of the WMG API
This is done be creating an exact copy of the files:
`v1.py` and `test_v1.py` and renaming them as `v2.py` and `test_v2.py`
respectively. The aim of this commit is to create an isolated fork
of the API code in which to develop the new functionality specified by
the WMG V2 requirements.