Add WMG V2 API capability to query without specifying tissues
This V2 version contains the functionality for the client to make queries
without specifying `tissue_ontology_term_id` in the request body. However,
the implementation is not entirely functional at this commit. This is not a
problem because the incomplete implementation is hidden from the main line
of code execution and therefore poses no risk due in the production environment.