refactor: Create a strawman V2 version of the WMG API (#4864)
* Create a V2 version of the WMG API

This is done be creating an exact copy of the files:
`v1.py` and `test_v1.py` and renaming them as `v2.py` and `test_v2.py`
respectively. The aim of this commit is to create an isolated fork
of the API code in which to develop the new functionality specified by
the WMG V2 requirements.

* Add WMG V2 API capability to query without specifying tissues

This V2 version contains the functionality for the client to make queries
without specifying `tissue_ontology_term_id` in the request body. However,
the implementation is not entirely functional at this commit. This is not a
problem because the incomplete implementation is hidden from the main line
of code execution and therefore poses no risk due in the production environment.