Make WMG V2 OpenApi spec more secure
Because `tissue_onotology_term_id` will not be specified in
the request body to `query` endpoint in WMG V2, this opens up the
possibility of injection attacks. Therefore, by specifying exactly
which properties should be included in the POST request body to an
API endpoint and rejecting all other properties, makes the API
more secure against injection attacks.