fix(wmg): do not subset anndata and save filter state instead (#5422)
Anndata is very memory inefficient. We skip filtering out rejected cells in the anndata object for this reason and simply record which cells must be filtered out in the Obs array. When it is time to construct the cubes by reading the Obs array from tiledb, we read the cell indices that must be filtered and perform the filtering on the pandas dataframe representation of the Obs array.

---------

Co-authored-by: Prathap Sridharan <psridharan@chanzuckerberg.com>