fix: do not assume latest DatasetVersion is active (#5425)
* fix: do not assume latest DatasetVersion is active

Change relevant db provider methods to strictly rely on datasets
associated with latest CollectionVersion as the 'active' DatasetVersion
for a given DatasetId. This fixes a bug introduced by #5355 which
allows rollback to previous DatasetVersions without deleting the
most-recently-created DatasetVersion row.

* add unit test

* add comments and modularize get_all_versions_for_dataset

This function is rendered obsolete by the changes in this commit series,
but the logic should stay around in case it is needed. Modularized.
Comment added about not currently being used

* docs: fix typo in method description

Co-authored-by: Trent Smith <1429913+Bento007@users.noreply.github.com>

* Update backend/layers/persistence/persistence.py

Co-authored-by: Trent Smith <1429913+Bento007@users.noreply.github.com>

---------

Co-authored-by: Trent Smith <1429913+Bento007@users.noreply.github.com>