undo errant addition of a datasets-bucket artifact to business test
Right now the public 'datasets' bucket assets are not tracked in the
DatasetArtifact table, as the 'artifacts' and 'cellxgene' bucket assets
are, so this commit undoes a recent feature commit that mistakenly added
a row for each type of asset for the 'datasets' assets to the
DatasetArtifact table. Update affected tests accordingly.