feat: Auto merge and deploy to staging branch (#5860)
Leverage existing workflow where a commit to `staging` branch creates
a deployment to the `stage` environment. So, whenever new commits to
`main` come in, we merge the commit to `staging` thus kicking off a
deployment to `stage` environment.

In addition, github action's concurrency group feature is
utilized to eliminate race conditions involving multiple builds and deployments
run at the same time. Here, we cancel scheduled builds and deployments in favor of the
newest build/deployment job.