fix: happy validating terraform that doesn't exist for single-cell
This PR fixes the cleanup action that is trying to validate the HCL in each of the environments. This is a newer feature that doesn't apply to this repository. I will commit a fix for the validating logic. Until now, roll back to the previous happy version before validations started.