fix: potential fix for staging deploy fail
Do not bother to use a 'reusable' workflow call in order to trigger a
staging deploy automatically after Push Tests (on main) succeed.
Instead, leverage existing 'on push' branches in the
build-images-and-create-deployment.yml workflow. This commit removes the
workflow call and adds 'staging' as a 'on push' branch for the latter.